+++ client/src/components/commons/DangerousActionModal.js
@@ -3,10 +3,10 @@ import React from "react";
 import {FaExclamation, FaQuestion} from "react-icons/all";
 import tinycolor from "tinycolor2";
 import {UiModal} from "ui";
-import {UiButton} from "ui/button";
+import {UiLoadableButton} from "ui/button";
 import {UiRow} from "ui/grid";
 
-const QuestionIcon = styled(FaQuestion)`
+export const QuestionIcon = styled(FaQuestion)`
   color: hsl(2, 95%, 66%);
   background-color: hsla(2, 95%, 66%, .1);
   height: 5rem;
@@ -18,12 +18,11 @@ const QuestionIcon = styled(FaQuestion)`
   margin-right: auto;
 `;
 
-const DangerousActionModal = ({id, isOpen, onHide, onAction, actionButtonName = "Delete", actionDescription}) => {
-    return <UiModal id={id} isOpen={isOpen} onHide={onHide} title={""} size={"sm"} className={"mx-0"}
-                    applyButton={<UiButton color={tinycolor("hsl(2, 95%, 66%)")} onClick={() => {
-                        onHide();
-                        onAction();
-                    }}><FaExclamation className={"move-top-1px"}/> {actionButtonName}</UiButton>}>
+const DangerousActionModal = ({id, isOpen, onHide, onAction, actionButtonName = "Delete", actionDescription, size = "sm"}) => {
+    return <UiModal id={id} isOpen={isOpen} onHide={onHide} title={""} size={size} className={"mx-0"}
+                    applyButton={<UiLoadableButton color={tinycolor("hsl(2, 95%, 66%)")} onClick={() => onAction().then(onHide)}>
+                        <FaExclamation className={"move-top-1px"}/> {actionButtonName}
+                    </UiLoadableButton>}>
         <UiRow centered className={"mt-3"}>
             <div className={"mb-2 px-4 text-center"}>
                 <QuestionIcon/>