+++ client/src/components/roadmap/SimpleIdeaCard.js
@@ -67,7 +67,7 @@ export const SimpleIdeaCard = ({ideaData}) => {
     return <IdeaContext.Provider value={{
         ideaData: idea, loaded: true, error: false,
         updateState: data => {
-            setIdea({...idea, data});
+            setIdea(data);
             history.replace({pathname: location.pathname, state: null});
         },
     }}>